Fix issue with os.stat reconstruction from bytes
The Google Drive API provides a stream that can be decompressed into an
os.stat object, but it uses the deprecated method '_make_stat_result'
from os that is not available anymore at least in Python 3.8. This
causes in Python 3 the following error because pickle does not know
how to handle it:

    AttributeError: "Can't get attribute '_make_stat_result' on
    <module 'os' from '/usr/lib/python3.8/os.py'>")

Because the behaviour of '_make_stat_result' can be reproduced with the
function 'stat_result' from os, this commit adds an intermediate step
for Python 3 in which the reference to '_make_stat_result' is replaced
with 'stat_result' before using pickle.